LPS-105380 Rename package to match class name
